Update V064
IMPLIMENTED CONFIG READING AND WRITING:

Game settings and configuration are now saved and stored so the player
can change them (to be implimented in a future update).

Game configuration values are now stored in the file: Config.Xodusse,
which is created when the game runs for the first time. This file can be
found at %LOCALAPPDATA%\Xodusse.

* Added scripts for config reading and writing.
* Settings determined in the obj_initialize create event.