Update V120
Display settings update!

*ADDED the persistent controller object obj_controller_menu, which is
used to handle events on the main menu screen. (Currently only used to
draw the game logo.)

*ADDED spr_logo_0 and spr_logo_1, (these are sprites for the game's
logo, but are split into 2 square images to decrease the texture swap
size, and also increase the pixel density). Menu sprites are placed in
the new sprite sub-folder (Menu).

*ADDED resolution settings 2 to 7. (2 = 1366x768, 3 = 1600x900, 4 =
1920x1080, 5 = 2650x1440, 6 = 3840x2160, 7 =
DISPLAY_HRESOLUTIONxDISPLAY_VRESOLUTION.) Setting 7 can only be
activated by toggling fullscreen mode, preventing the player from
scaling the window manually. The game will also detect the maximum 16:9
screen ratio it can set itself to, depending on the resolution of the
current monitor. The game also reads and writes the current resolution
from and to the Config.Xodusse file located in %appdata%.

*ADDED resolution events to the obj_button_medium object when the player
switches the resolution. This button can only be pressed when not using
fullscreen.

*MODIFIED obj_controller_gui so that the W and H variables used in the
draw event use 2 new variables called GUIW and GUIH, which represent the
size of the GUI drawing region. (Initialized in the create event of
obj_initialize).

*MODIFIED the old spr_button_large sprite, so now it is not a button at
all, just a small crosshair to determine where the button is placed in
the room. (This button is now an animated shape which is drawn onto the
screen.)

*UPDATED obj_controller_debugger version number, and Global Game
Settings version number.

In-game HUD now also scales in size to the current game resolution, so
that it remains the same size across all different monitor sizes.
(Scaling issues with selection of inventory items have also been fixed.)