Update V125
Resolution scaling update.

The game now scales proportionately to fit any monitor size that is not
16:9 (although currently buggy).

*CREATED obj_init_display in the initialization folder which handles
display update events.

*DISABLED views in every room because these are now enabled
automatically for any new room that is created. (This is most likely why
the main menu stretches on monitors with unusual aspect ratios.)

* Upon modifying the game resolution in game, this also resizes the
application surface and centers the window. (called in
obj_button_medium.)

* Updated the debug screen version number.
* Updated the Global Game Settings version number.