Update V144
IMPLEMENTED PAUSE SCREEN. 👍

*CREATED the object obj_button_pause (the pause button has to have
different behaviour as opposed to a regular button, due to the fact that
it can actually change position inside the game room).

*UPDATED obj_initialize by adding a new variable 'VIEW_ANGLE', which
stores the current angle of the camera in the game. (Updated every step
by using the view_angle[0] variable, of which the result is the modulus
of this value by 360.

*UPDATED obj_controller_gui, such that it now draws a pause screen and
darkens the game room. (This is rendered as two partially opaque
triangles and has it's position calculated according to the view size
and rotation using VIEW_ANGLE). The pause screen is also drawn using a
standard 'Draw' event, and NOT 'Draw GUI'. (This is because the button
objects do not function correctly whilst using 'Draw GUI'.)

*MODIFIED obj_player and obj_enemy, to account for pausing the game. If
the global variable 'PAUSED' is set to true, this halts all events for
the given objects.

*MODIFIED obj_controller_keys so that it now pauses the game and creates
instances of obj_button_pause as the game is paused. Obj_button_pause
instances then check each step if the game is still paused, and if not
they are destroyed.

*UPDATED both the debug screen version and Global Game Settings build
number.