Update V174
Made some huge changes...

*ADDED a new cursor sprite which animates upon hovering over buttons.
(This was increcibly complicated because the view rotation wasn't
accounted for when drawing the sprite to the GUI layer, this is fixed
however).

*ADDED new commands to the console for debugging, these include /god
(makes you invincible), /killall (kills all current aliens in existance,
resets the wave counter and starts a new round), /alienmax [n]
(specifies the maximum amount of aliens that can spawn at once, (set
this to 0 to stop the aliens spawning completely)).

*ADDED the object obj_controller_cursor for drawing the cursor sprite.

*ADDED 2 global variables into obj_initialize - Create.
FRAME_DELTA_OVERRIDE is used to override the FRAME_DELTA being 0 whilst
the game is paused. (It is believed that this was the bug on the pause
screen where a blackscreen occured, this is known to be fixed).
CHEAT_GODMODE is used for toggling god mode.

*ADDED (TEMPORARILY) a blinking light in the game room using a dsin()
function to make it pulse, (this shows the location of the backpack in
the level for the player to pick up). This light is disabled when the
player picks up the backpack (obj_item). This triggers custom event 0 in
obj_normal to disable the light).

*ADDED the scripts scr_command_alienmax, _god, _killall. (Modified _help
to display the new commands).

*MODIFIED obj_controller_debugger so that it displays if god mode is
enabled.

*MODIFIED obj_button_large,pause and small to account for the cursor
target animation.

*MODIFIED obj_player to support god mode.

*MODIFIED scr_lighting_draw_level so that it expands the level texture
by 10x.

*FIXED a small bug with dropping items into the level from the
inventory, (this originally crashed the game).

*DUPLICATED rm_game_0 as a back-up and modified the wall tiles so that
they are not the standard square sprites anymore through object creation
code.

*REMOVED the show debug bar from obj_initialize.

*UPDATED the Global Game Settings version number.