Update V194
*Implimented a new alien, the Waggle, and the evolved form of it, the
Writhe...*
Both the Writhe and the Waggle currently use the same AI as Alien_1
(which I now want to call a Groan). However, the Writhe slithers across
terrain using a worm-like spine. It also has glowing eyes which is
pretty creepy.

Created new Alien child types for each different alien with the parent
being obj_enemy.

*ADDED various diffuse, she and normal sprites for the new alien
variant.
*ADDED new objects and tail objects for the new alien variant.

*MODIFIED the wave controller so that there is a 20% chance of spawning
a Wraggle, and if this 20% chance occurs, there is then a 40% chance of
spawning a Writhe.